# api/swagger.yaml
openapi: 3.0.3
info:
  title: Jobot API
  description: |
    REST API для Telegram бота Jobot - системы поиска работы и размещения вакансий.
    
    ## Возможности
    - Управление пользователями (сотрудники и работодатели)
    - Создание и управление профилями
    - Публикация вакансий
    - Загрузка резюме
    - Система лайков (matching)
    
  version: 2.0.0
  contact:
    name: Jobot Team
    url: https://github.com/sashasych/jobot
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.jobot.com
    description: Production server
...
paths:
...
  /api/employees:
    post:
      tags:
        - employees
      summary: Создать профиль сотрудника
      description: Создание профиля соискателя
      operationId: createEmployee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeCreateRequest'
      responses:
        '201':
          description: Employee created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
        '400':
          description: Bad request
  /api/employees/{employeeId}:
    get:
      tags:
        - employees
      summary: Получить сотрудника
      operationId: getEmployeeById
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponse'
        '404':
          description: Employee not found
    
    put:
      tags:
        - employees
      summary: Обновить сотрудника
      operationId: updateEmployee
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeUpdateRequest'
      responses:
        '200':
          description: Employee updated successfully
        '404':
          description: Employee not found
    
    delete:
      tags:
        - employees
      summary: Удалить сотрудника
      operationId: deleteEmployee
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Employee deleted successfully
        '404':
          description: Employee not found
...
  /api/employees/{employeeId}/reactions:
    get:
      tags:
        - employees
      summary: Получить реакции сотрудника
      description: Получение всех реакций (лайков) сотрудника
      operationId: getEmployeeReactions
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeReactionListResponse'
  /api/employers:
    post:
      tags:
        - employers
      summary: Создать профиль работодателя
      operationId: createEmployer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployerCreateRequest'
      responses:
        '201':
          description: Employer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployerResponse'
  /api/employers/{employerId}:
    get:
      tags:
        - employers
      summary: Получить работодателя
      operationId: getEmployerById
      parameters:
        - name: employerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployerResponse'
    
    put:
      tags:
        - employers
      summary: Обновить работодателя
      operationId: updateEmployer
      parameters:
        - name: employerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployerUpdateRequest'
      responses:
        '200':
          description: Employer updated successfully
    
    delete:
      tags:
        - employers
      summary: Удалить работодателя
      operationId: deleteEmployer
      parameters:
        - name: employerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Employer deleted successfully
...
  /api/reactions:
    post:
      tags:
        - reactions
      summary: Создать реакцию
      description: Поставить лайк на вакансию
      operationId: createReaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReactionCreateRequest'
      responses:
        '201':
          description: Reaction created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReactionResponse'
components:
  schemas:
...
    EmployeeCreateRequest:
      type: object
      required:
        - user_id
        - tags
      properties:
        user_id:
          type: string
          format: uuid
        tags:
          type: array
          items:
            type: string
          example: ["golang", "postgresql", "docker"]
    EmployeeUpdateRequest:
      type: object
      properties:
        tags:
          type: array
          items:
            type: string
    EmployeeResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            employee_id:
              type: string
              format: uuid
            user_id:
              type: string
              format: uuid
            tags:
              type: array
              items:
                type: string
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    EmployerCreateRequest:
      type: object
      required:
        - user_id
        - company_name
        - company_description
        - company_location
        - company_size
      properties:
        user_id:
          type: string
          format: uuid
        company_name:
          type: string
          example: TechCorp Inc
        company_description:
          type: string
          example: Leading technology company
        company_website:
          type: string
          example: https://techcorp.com
        company_location:
          type: string
          example: Москва, Россия
        company_size:
          type: string
          example: 51-200
    EmployerUpdateRequest:
      type: object
      properties:
        company_name:
          type: string
        company_description:
          type: string
        company_website:
          type: string
        company_location:
          type: string
        company_size:
          type: string
    EmployerResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            employer_id:
              type: string
              format: uuid
            user_id:
              type: string
              format: uuid
            company_name:
              type: string
            company_description:
              type: string
            company_website:
              type: string
            company_location:
              type: string
            company_size:
              type: string
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
...
    ReactionCreateRequest:
      type: object
      required:
        - employee_id
        - vacansie_id
        - reaction
      properties:
        employee_id:
          type: string
          format: uuid
        vacansie_id:
          type: string
          format: uuid
        reaction:
          type: string
          enum: [like, dislike]
    ReactionResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            reaction_id:
              type: string
              format: uuid
            employee_id:
              type: string
              format: uuid
            vacansie_id:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
    EmployeeReactionListResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            reactions_ids:
              type: array
              items:
                type: string
                format: uuid
            employee_id:
              type: string
              format: uuid
...